+++ test/test_images.c
@@ -211,8 +211,8 @@ int main(void)
 
 	if (test_data_dir_create(g_data_dir, sizeof(g_data_dir)) != 0)
 		return 1;
-	snprintf(g_pkg_state_dir, sizeof(g_pkg_state_dir), "%s/pkg", g_data_dir);
-	snprintf(g_images_dir, sizeof(g_images_dir), "%s/images", g_data_dir);
+	snprintf(g_pkg_state_dir, sizeof(g_pkg_state_dir), "%s/rebuildable/pkg", g_data_dir);
+	snprintf(g_images_dir, sizeof(g_images_dir), "%s/rebuildable/images", g_data_dir);
 
 	reset_state();
 	run_cmd("mkdir -p '%s/recipes'", g_pkg_state_dir);